Se agrego validacion de marca, pero aun no se ha probado en el postman

+++ src/main/java/com/edu/mx/inte5A/Marca/Model/MarcaRepository.java
@@ -8,7 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MarcaRepository extends JpaRepository<Marca, Integer> {
+public interface MarcaRepository extends JpaRepository<Marca, Long> {
     Optional<Marca> findByNombre(String nombre);
     List<Marca> findAllByStatusIsTrue();
+
 }
